ACTIVATED THE GATSBY-PLUGIN-OFFLINE FOR WORKBOX AND SERVICE WORKER, ADDED SERVICE WORKER REGISTRATION TO INDEX.JS
Activated the service worker plugin by running gatsby build which installed workbox tooling and a service worker (sw.js) file. I then registered the serice worker in the index.js file. Tested the service worker and it is basically working offline now. It is setup to cache the pages a user goes to when on the network and have those pages available when they go offline. However we still need to setup precaching so some pages are automatically cached whether on the network or not, and also updat the 404/ offline page.